cm 2019-03-15 01

+++ source/LiteDbExplorer/App.xaml.cs
@@ -6,10 +6,10 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Windows;
+using System.Windows.Threading;
 using LiteDbExplorer.Controls;
+using LiteDbExplorer.Framework.Windows;
 using LiteDbExplorer.Presentation;
-using Onova;
-using Onova.Services;
 
 namespace LiteDbExplorer
 {
@@ -21,10 +21,19 @@ namespace LiteDbExplorer
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
         private readonly string _instanceMuxet = "LiteDBExplorerInstaceMutex";
         private Mutex _appMutex;
+        private bool _errorNotified;
 
         public bool OriginalInstance { get; private set; }
         
         public static Settings Settings => Settings.Current;
+
+        public App()
+        {
+            AppDomain.CurrentDomain.UnhandledException += OnCurrentDomainUnhandledException;
+            DispatcherUnhandledException += OnDispatcherUnhandledException;
+
+            Config.ConfigureLogger();
+        }
 
         protected override void OnStartup(StartupEventArgs e)
         {
@@ -43,11 +52,6 @@ namespace LiteDbExplorer
 
         private void Application_Startup(object sender, StartupEventArgs e)
         {
-#if !DEBUG
-            AppDomain.CurrentDomain.UnhandledException += CurrentDomain_UnhandledException;
-#endif
-            Config.ConfigureLogger();
-
             ThemeManager.SetColorTheme(Settings.ColorTheme);
 
 
@@ -72,7 +76,7 @@ namespace LiteDbExplorer
                 OriginalInstance = true;
             }
         }
-
+        
         private void Settings_PropertyChanged(object sender, PropertyChangedEventArgs e)
         {
             switch (e.PropertyName)
@@ -95,16 +99,68 @@ namespace LiteDbExplorer
             _appMutex?.ReleaseMutex();
         }
 
-        private void CurrentDomain_UnhandledException(object sender, UnhandledExceptionEventArgs e)
+        private void OnCurrentDomainUnhandledException(object sender, UnhandledExceptionEventArgs e)
         {
             var log = LogManager.Configuration.FindTargetByName("file") as FileTarget;
             Logger.Error((Exception) e.ExceptionObject, "Unhandled exception: ");
-            MessageBox.Show(
-                string.Format(
-                    "Unhandled exception occured.\nAdditional information written into: {0}\n\nApplication will shutdown.",
-                    log.FileName),
-                "Unhandled Exception", MessageBoxButton.OK, MessageBoxImage.Error);
-            Process.GetCurrentProcess().Kill();
+
+            var message = "Unhandled exception occured.\n";
+            if (log != null)
+            {
+                message += $"\nAdditional information written into: {log.FileName}.\n";
+            }
+            if (e.IsTerminating)
+            {
+                message += "\nApplication will shutdown.\n";
+            }
+            
+            ShowError((Exception) e.ExceptionObject, message, "Unhandled Exception");
+            
+            if (e.IsTerminating)
+            {
+                Process.GetCurrentProcess().Kill();
+            }
+        }
+
+        private void OnDispatcherUnhandledException(object sender, DispatcherUnhandledExceptionEventArgs e)
+        {
+            var log = LogManager.Configuration.FindTargetByName("file") as FileTarget;
+            Logger.Error(e.Exception, "An unhandled exception occurred");
+
+            var message = "Unhandled exception occured.\n";
+            if (log != null)
+            {
+                message += $"\nAdditional information written into: {log.FileName}.\n";
+            }
+            if (!e.Handled)
+            {
+                message += "\nApplication will shutdown.\n";
+            }
+
+            ShowError(e.Exception, message, "Unhandled Exception");
+        }
+
+        public void ShowError(Exception exception, string message, string caption = "")
+        {
+            if(_errorNotified)
+            {
+                return;
+            }
+            
+            _errorNotified = true;
+
+            var exceptionViewer = new ExceptionViewer(message, exception);
+            var baseDialogWindow = new BaseDialogWindow
+            {
+                Title = string.IsNullOrEmpty(caption) ? "Error" : caption,
+                Content = exceptionViewer,
+                WindowStartupLocation = WindowStartupLocation.CenterScreen,
+                ResizeMode = ResizeMode.NoResize,
+                IsMinButtonEnabled = false
+            };
+            baseDialogWindow.ShowDialog();
+
+            // MessageBox.Show(message, caption, MessageBoxButton.OK, MessageBoxImage.Error);
         }
     }
 }
